First step to loading and saving rules
This commit only saves parts of the Game objects. But it shows that it
works.

starting # with '#' will be ignored, and an empty message aborts the
commit.  # # Date: Fri Apr 27 23:49:35 2018 +0200 # # On branch master
push" to publish your local commits) # # Changes to be committed: #
modified: Game.cs # modified: Program.cs # modified:
SurvivalAIGame.csproj #